feat: progress with new tag form dialog
