+++ src/pages/monitoring/outlets/tagsAdmin/layout/tableContent.tsx
@@ -1,15 +1,16 @@
 import type { ODataColumn } from "@appTypes/odata";
 import { TagEntryShort } from "pages/monitoring/types/odataResponse";
+import { TagForm } from "pages/monitoring/outlets/tagsAdmin/TagFormBtn";
 
 export const tableContent: ODataColumn<TagEntryShort>[] = [
   { name: "Categoría", source: "categoryName", type: "text", sortBy: true },
   { name: "Nombre", source: "name", type: "text", sortBy: true },
   { name: "Url", source: "url", type: "text" },
-  // {
-  //   name: "Acciones",
-  //   source: "id",
-  //   type: "action",
-  //   actions: ShowLogDetailsButton,
-  //   label: "Detalles",
-  // },
+  {
+    name: "Acciones",
+    source: "id",
+    type: "action",
+    actions: TagForm,
+    label: "Detalles",
+  },
 ];
